+++ tests/single_broker/.devel/run.sh
@@ -13,7 +13,7 @@ export PROJECT_HOME=$projectHome
 
 ansibleSolaceTests=(
   "setup"
-  # "solace_client_profile"
+  "solace_client_profile"
   "solace_replay"
   "solace_cert_authority"
   "solace_get_list"
@@ -38,7 +38,7 @@ export ANSIBLE_SOLACE_TESTS="${ansibleSolaceTests[*]}"
   export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:latest"
   # export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:9.9.0.23"
   # export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:9.8.0.12"
-  export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:9.7.0.42"
+  # export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:9.7.0.42"
   # export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:9.6.0.46"
   # export BROKER_DOCKER_IMAGE="solace/solace-pubsub-standard:9.3.1.28"
   export BROKER_TYPE="local"
@@ -46,10 +46,10 @@ export ANSIBLE_SOLACE_TESTS="${ansibleSolaceTests[*]}"
   export BROKER_DOCKER_COMPOSE_FILE="$projectHome/test-runner/files/PubSubStandard_singleNode.yml"
   export INVENTORY_FILE=$LOCAL_BROKER_INVENTORY_FILE
 
-# # solace cloud broker
-#   export BROKER_TYPE="solace_cloud"
-#   export INVENTORY_FILE="$projectHome/test-runner/files/solace-cloud-account.inventory.yml"
-#   export SOLACE_CLOUD_API_TOKEN=$SOLACE_CLOUD_API_TOKEN_ALL_PERMISSIONS
+# solace cloud broker
+  export BROKER_TYPE="solace_cloud"
+  export INVENTORY_FILE="$projectHome/test-runner/files/solace-cloud-account.inventory.yml"
+  export SOLACE_CLOUD_API_TOKEN=$SOLACE_CLOUD_API_TOKEN_ALL_PERMISSIONS
 
 export CLEAN_WORKING_DIR=False