+++ tests/single_broker/_run.sh
@@ -18,6 +18,7 @@ source $PROJECT_HOME/.lib/functions.sh
   if [ -z "$ANSIBLE_SOLACE_TESTS" ]; then
     export ANSIBLE_SOLACE_TESTS=(
       "setup"
+      "solace_client_profile"
       "solace_replay"
       "solace_get_list"
       "solace_service_auth"
@@ -26,7 +27,6 @@ source $PROJECT_HOME/.lib/functions.sh
       "solace_oauth"
       "solace_facts"
       "solace_vpn"
-      # "solace_client_profile" - broken solace cloud api: 14/July/2021
       "solace_acl_profile"
       "solace_rdp"
       "solace_cert_authority"